Move form to FormComponent
