+++ src/views/pages/authentication/sign-up/SignUpPage.vue
@@ -1,12 +1,14 @@
 <script setup lang="ts">
     import { defineAsyncComponent, reactive } from 'vue';
-    import APIService from '@/core/services/api-service';
     import UserInterface from '@/interfaces/UserInterface';
-    import validationHandler from '@/core/utilities/validation/validation-hendler';
 
-    const InputGroup = defineAsyncComponent(() => import('@/components/input-components/InputGroup.vue'));
+    const FormComponent = defineAsyncComponent(() => import('@/components/form-components/FormComponent.vue'));
 
     const data = {
+        config: {
+            URL: '/users',
+            API: '/users'
+        },
         formData: {
             'firstName': '',
             'lastName': '',
@@ -28,31 +30,8 @@
         ]
     };
 
-    const formData = reactive<UserInterface>({ ...data.formData });
+    const formData = reactive<UserInterface>(data.formData);
 
-    const validationErrors = reactive(
-        Object.fromEntries(Object.keys(formData).map(key => [key, '']))
-    );
-
-    const handleSubmit = async (): Promise<void> => {
-        const formElement = document.querySelector('form');
-        const [isValid, errors] = validationHandler.validateFormData(formElement);
-
-        for (const key of Object.keys(validationErrors)) {
-            validationErrors[key] = '';
-        }
-
-        if (isValid) {
-            await APIService.store('/users', formData);
-            resetFormData();
-        } else {
-            Object.assign(validationErrors, errors);
-        }
-    };
-
-    const resetFormData = (): void => {
-        Object.assign(formData, data.formData);
-    };
 </script>
 
 <template>
@@ -61,16 +40,12 @@
             <div class="bg-white p-6 rounded-lg shadow-md">
                 <h2 class="mt-5 text-center text-2xl font-bold leading-9 tracking-tight text-gray-900">Sign up</h2>
                 <div class="mt-10">
-                    <form class="space-y-4" @submit.prevent="handleSubmit">
-                        <InputGroup
-                            :form-data="formData"
-                            :inputs="data.inputs"
-                            :validation-errors="validationErrors"
-                        />
-                        <div class="pt-4">
-                            <button type="submit" class="flex w-full justify-center rounded-md bg-primary px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600">Sign up</button>
-                        </div>
-                    </form>
+                    <FormComponent
+                        class="space-y-4"
+                        :config="data.config"
+                        :data="formData"
+                        :inputs="data.inputs"
+                    />
                 </div>
             </div>
         </div>
